refactor(inspection-edit): extract loader + action into server modules (3003->2487)
Batch 1 of splitting the inspection-edit route. Moves the loader (79 lines) and
the 20-intent action dispatcher (434 lines) VERBATIM into co-located
.server.ts modules, re-exported so the route's RR v7 module exports (loader,
action, meta, shouldRevalidate, default component) are unchanged.

Behavior-preserving: reviewer diffed loader/action/component bodies against the
prior commit -> byte-identical. Moved-only imports relocated; shared helpers
kept route-local and imported, not duplicated. Gates: type-check 0, lint 0,
test:unit 1937|8, test:web 516, db:check clean.
